Created ES6 modules and used import/export functions
